refactor: apply suggested file format changes
