+++ library/lib/edges/index.ts
@@ -1,9 +1,2 @@
 export * from "./types"
-// export * from "./AggregationEdge"
-// export * from "./BiAssociationEdge"
-// export * from "./CompositionEdge"
-// export * from "./DependencyEdge"
-// export * from "./InheritanceEdge"
-// export * from "./RealizationEdge"
-// export * from "./UniAssociationEdge"
 export * from "./EdgeProps"
