+++ library/lib/hooks/index.ts
@@ -7,3 +7,4 @@ export * from "./useConnect"
 export * from "./useNodeDragStop"
 export * from "./useHandleOnResize"
 export * from "./useEdges"
+export * from "./useEdgeToolbar"
