+++ library/lib/hooks/useConnect.ts
@@ -12,12 +12,6 @@ export const useConnect = () => {
         id: generateUUID(),
         type: "ClassBidirectional", // Ensure this type matches your custom edge in diagramEdgeTypes
         selected: true,
-        data: {
-          sourceRole: "", // Default empty string for source role
-          sourceMultiplicity: "", // Default empty string for source multiplicity
-          targetRole: "", // Default empty string for target role
-          targetMultiplicity: "", // Default empty string for target multiplicity
-        },
       }
       setEdges((edges) => addEdge(newEdge, edges))
     },
